pyproject.toml
[project]
name = "bds-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.43",
    "ipykernel>=6.30.1",
]
